Create db tables
Please make sure to comment out the 'Drop" statements prior to running these statements. 
In order to create these, you will first have to create a .db file and store it somewhere in a folder of your choice. This can be done by simply opening a notepad and save the file type as 'All types' and then give the file a name and include the .db extension
After the creation of that file, navigate to PyCharm and connect data source. 
After you connect a data source, specify the path of the newly created .db file. Test the connection and ensure there are no errors. 
Then proceed to apply these changes. Then run the below statements

These statements will create a db with the following tables;
- users,
- bugs
- changes
- mentions
- comments

Note that the Drop table statements will drop the respective tables mentioned.